Simplify the SelectChildren operation.
We remove `traverse_up_to_node_which_matters()` because if we have multiple
selections (such as from a previous invocation of SelectChildren) that ensures
we select the children of the node containing all the selections, rather than
the children of each selection.

We remove the code to preserve selected nodes with no children. Those
are typically keywords and delimiting punctuation; if the user invokes
SelectChildren on an array literal, they probably want to work with each item
rather than with the brackets and commas. Also, this change makes SelectNode and
SelectChildren into opposites. Previously, after invoking SelectChildren any
number of times, invoking SelectNode once would return to the initial selection,
because some direct child of the initial node would still be selected. Now if
you SelectChildren three times, you'll probably have to SelectNode three times
to get back, which makes more sense.